separate file into interface.go.
